+++ packages/gatsby-source-contentful/src/create-schema-customization.js
@@ -70,7 +70,7 @@ async function getContentTypesFromContentful({
 }
 
 export async function createSchemaCustomization(
-  { schema, actions, reporter, cache },
+  { schema, actions, store, reporter, cache },
   pluginOptions
 ) {
   const { createTypes } = actions
@@ -166,6 +166,7 @@ export async function createSchemaCustomization(
       {
         schema,
         actions,
+        store,
       }
     )
   )