+++ packages/gatsby-source-contentful/src/download-contentful-assets.js
@@ -78,11 +78,9 @@ export async function downloadContentfulAssets(gatsbyFunctions) {
       if (!fileNodeID) {
         const fileNode = await createRemoteFileNode({
           url,
-          store,
           cache,
           createNode,
           createNodeId,
-          reporter,
         })
 
         if (fileNode) {