+++ packages/gatsby/cache-dir/production-app.js
@@ -1,4 +1,3 @@
-/* global HAS_REACT_18 */
 import { apiRunner, apiRunnerAsync } from "./api-runner-browser"
 import React from "react"
 import { Router, navigate, Location, BaseContext } from "@gatsbyjs/reach-router"
@@ -24,23 +23,13 @@ import stripPrefix from "./strip-prefix"
 
 // Generated during bootstrap
 import matchPaths from "$virtual/match-paths.json"
+import { reactDOMUtils } from "./react-dom-utils"
 
 const loader = new ProdLoader(asyncRequires, matchPaths, window.pageData)
 setLoader(loader)
 loader.setApiRunner(apiRunner)
 
-let reactHydrate
-let reactRender
-if (HAS_REACT_18) {
-  const reactDomClient = require(`react-dom/client`)
-  reactRender = (Component, el) =>
-    reactDomClient.createRoot(el).render(Component)
-  reactHydrate = (Component, el) => reactDomClient.hydrateRoot(el, Component)
-} else {
-  const reactDomClient = require(`react-dom`)
-  reactRender = reactDomClient.render
-  reactHydrate = reactDomClient.hydrate
-}
+const { render, hydrate } = reactDOMUtils()
 
 window.asyncRequires = asyncRequires
 window.___emitter = emitter
@@ -266,9 +255,9 @@ apiRunnerAsync(`onClientEntry`).then(() => {
 
     // Client only pages have any empty body so we just do a normal
     // render to avoid React complaining about hydration mis-matches.
-    let defaultRenderer = reactRender
+    let defaultRenderer = render
     if (focusEl && focusEl.children.length) {
-      defaultRenderer = reactHydrate
+      defaultRenderer = hydrate
     }
 
     const renderer = apiRunner(