+++ packages/gatsby-remark-autolink-headers/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.16.0](https://github.com/gatsbyjs/gatsby/commits/gatsby-remark-autolink-headers@5.16.0/packages/gatsby-remark-autolink-headers) (2022-06-07)
+
+[🧾 Release notes](https://www.gatsbyjs.com/docs/reference/release-notes/v4.16)
+
+**Note:** Version bump only for package gatsby-remark-autolink-headers
+
+### [5.15.1](https://github.com/gatsbyjs/gatsby/commits/gatsby-remark-autolink-headers@5.15.1/packages/gatsby-remark-autolink-headers) (2022-06-01)
+
+**Note:** Version bump only for package gatsby-remark-autolink-headers
+
+## [5.15.0](https://github.com/gatsbyjs/gatsby/commits/gatsby-remark-autolink-headers@5.15.0/packages/gatsby-remark-autolink-headers) (2022-05-24)
+
+[🧾 Release notes](https://www.gatsbyjs.com/docs/reference/release-notes/v4.15)
+
+**Note:** Version bump only for package gatsby-remark-autolink-headers
+
 ## [5.14.0](https://github.com/gatsbyjs/gatsby/commits/gatsby-remark-autolink-headers@5.14.0/packages/gatsby-remark-autolink-headers) (2022-05-10)
 
 [🧾 Release notes](https://www.gatsbyjs.com/docs/reference/release-notes/v4.14)