+++ packages/gatsby-transformer-documentationjs/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.16.0](https://github.com/gatsbyjs/gatsby/commits/gatsby-transformer-documentationjs@6.16.0/packages/gatsby-transformer-documentationjs) (2022-06-07)
+
+[🧾 Release notes](https://www.gatsbyjs.com/docs/reference/release-notes/v4.16)
+
+**Note:** Version bump only for package gatsby-transformer-documentationjs
+
+## [6.15.0](https://github.com/gatsbyjs/gatsby/commits/gatsby-transformer-documentationjs@6.15.0/packages/gatsby-transformer-documentationjs) (2022-05-24)
+
+[🧾 Release notes](https://www.gatsbyjs.com/docs/reference/release-notes/v4.15)
+
+**Note:** Version bump only for package gatsby-transformer-documentationjs
+
 ## [6.14.0](https://github.com/gatsbyjs/gatsby/commits/gatsby-transformer-documentationjs@6.14.0/packages/gatsby-transformer-documentationjs) (2022-05-10)
 
 [🧾 Release notes](https://www.gatsbyjs.com/docs/reference/release-notes/v4.14)